requirements.txt
Django==1.11.6
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
pytz==2017.2